Moved back Library files from Client folder to UserAgent.
